requirements.txt
streamlit==1.28.1
ultralytics==8.0.196
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
scikit-learn==1.3.0
torch>=1.13.0
torchvision>=0.14.0
pandas==2.0.3